+++ rmm/src/org/dragon/rmm/model/ResUser.java
@@ -2,15 +2,15 @@ package org.dragon.rmm.model;
 
 public class ResUser extends InfoUserLogin {
 
-	public int userid;
-	public String nickname;
-	public String address;
+    public long userid;
+    public String nickname;
+    public String address;
 
-	public ResUser(String username, String password, int userid, String nickname, String address) {
-		super(username, password);
-		this.userid = userid;
-		this.nickname = nickname;
-		this.address = address;
-	}
+    public ResUser(String username, String password, long userid, String nickname, String address) {
+        super(username, password);
+        this.userid = userid;
+        this.nickname = nickname;
+        this.address = address;
+    }
 
 }
